# openapi/ai-capabilities.yaml
openapi: "3.0.3"
info:
  title: AI 能力 API
  description: |
    AI 能力模块，包含：
    - AI 核心（对话、版本、行为、同步）
    - AI 健康检查
    - AI 监控
    - 插件市场
    - RAG 知识库
    - LLM Provider
    - AI 协作
    - 提示词模板
    - Agent Workflow
    - 高级功能
    - Provider 管理
    - AI 模型
    - 文档解析
  version: "1.0.0"
servers:
  - url: http://localhost:3001
    description: 开发环境
...
paths:
...
  /api/ai/prompts/{id}/test:
    post:
      tags: [ai-prompts]
      summary: 测试提示词
      operationId: testPrompt
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                variables:
                  type: object
                  description: "模板变量值"
      responses:
        "200":
          description: 测试成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      rendered:
                        type: string
                        description: "渲染后的提示词"
                      result:
                        type: string
                        description: "AI 响应"
...
  /api/providers/{id}/models:
    get:
      tags: [ai-providers]
      summary: Provider 下的模型列表
      operationId: listProviderModels
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        type:
                          type: string

  # ==================== AI 模型 ====================
  /api/models:
    get:
      tags: [ai-models]
      summary: 模型列表
      operationId: listAiModels
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: providerId
          in: query
          description: "Provider ID"
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
    post:
      tags: [ai-models]
      summary: 创建模型
      operationId: createAiModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - providerId
              properties:
                name:
                  type: string
                providerId:
                  type: string
                type:
                  type: string
                  enum: ["chat", "embedding", "image"]
                description:
                  type: string
                config:
                  type: object
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/models/{id}:
    get:
      tags: [ai-models]
      summary: 模型详情
      operationId: getAiModel
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    put:
      tags: [ai-models]
      summary: 更新模型
      operationId: updateAiModel
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                config:
                  type: object
                status:
                  type: string
                  enum: ["active", "disabled"]
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    delete:
      tags: [ai-models]
      summary: 删除模型
      operationId: deleteAiModel
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/models/{id}/test:
    post:
      tags: [ai-models]
      summary: 测试模型连通性
      operationId: testAiModel
      parameters:
        - $ref: "#/components/parameters/ObjectIdParam"
      responses:
        "200":
          description: 测试成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: ["ok", "error"]
                      message:
                        type: string
                      latency:
                        type: number
                        description: "响应时间（ms）"

  # ==================== 文档解析 ====================
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: object
          properties:
            message:
              type: string
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            totalPages:
              type: integer
...
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: 每页数量
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ObjectIdParam:
      name: id
      in: path
      required: true
      description: "资源 ID（MongoDB ObjectId）"
      schema:
        type: string
        pattern: "^[0-9a-fA-F]{24}$"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    workflowKeyAuth:
      type: apiKey
      in: header
      name: X-Workflow-Key